+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/config/RuleRegistry.java
@@ -28,8 +28,8 @@ import io.shardingsphere.core.executor.ShardingExecuteEngine;
 import io.shardingsphere.core.metadata.ShardingMetaData;
 import io.shardingsphere.core.orche.config.ProxyBasicRule;
 import io.shardingsphere.core.event.orche.config.ProxyConfigurationEventBusEvent;
-import io.shardingsphere.core.orche.eventbus.state.CircuitStateEventBusEvent;
-import io.shardingsphere.core.orche.eventbus.state.DisabledStateEventBusEvent;
+import io.shardingsphere.core.event.orche.state.CircuitStateEventBusEvent;
+import io.shardingsphere.core.event.orche.state.DisabledStateEventBusEvent;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.core.rule.MasterSlaveRule;
 import io.shardingsphere.core.rule.ProxyAuthority;
